refactor: remove connector session handoff
