+++ turbo/apps/api/src/signals/routes/connector-oauth-route-state.ts
@@ -1,9 +1,6 @@
-import { env } from "../../lib/env";
-
-export const CONNECTOR_OAUTH_STATE_COOKIE_NAME = "connector_oauth_state";
-export const CONNECTOR_OAUTH_SESSION_COOKIE_NAME = "connector_oauth_session";
-export const CONNECTOR_OAUTH_PKCE_COOKIE_NAME = "connector_oauth_pkce";
-export const CONNECTOR_OAUTH_CONTEXT_COOKIE_NAME = "connector_oauth_context";
+const CONNECTOR_OAUTH_STATE_COOKIE_NAME = "connector_oauth_state";
+const CONNECTOR_OAUTH_PKCE_COOKIE_NAME = "connector_oauth_pkce";
+const CONNECTOR_OAUTH_CONTEXT_COOKIE_NAME = "connector_oauth_context";
 export const CONNECTOR_OAUTH_COOKIE_MAX_AGE_SECONDS = 15 * 60;
 
 const CONNECTOR_OAUTH_REDIRECT_STATUS = 307;
@@ -14,24 +11,6 @@ export function generateConnectorOAuthState(): string {
   return Array.from(array, (byte) => {
     return byte.toString(16).padStart(2, "0");
   }).join("");
-}
-
-export function buildConnectorOAuthCookieHeader(
-  name: string,
-  value: string,
-  maxAge: number,
-): string {
-  const parts = [
-    `${name}=${encodeURIComponent(value)}`,
-    `Max-Age=${maxAge}`,
-    "Path=/",
-    "HttpOnly",
-    "SameSite=Lax",
-  ];
-  if (env("ENV") === "production") {
-    parts.push("Secure");
-  }
-  return parts.join("; ");
 }
 
 function buildDeleteConnectorOAuthCookieHeader(name: string): string {
@@ -49,10 +28,6 @@ export function clearConnectorOAuthCookies(response: Response): void {
   response.headers.append(
     "Set-Cookie",
     buildDeleteConnectorOAuthCookieHeader(CONNECTOR_OAUTH_STATE_COOKIE_NAME),
-  );
-  response.headers.append(
-    "Set-Cookie",
-    buildDeleteConnectorOAuthCookieHeader(CONNECTOR_OAUTH_SESSION_COOKIE_NAME),
   );
   response.headers.append(
     "Set-Cookie",
