+++ turbo/apps/api/src/signals/routes/zero-connectors.ts
@@ -1,45 +1,30 @@
-import { randomInt } from "node:crypto";
-
 import { command, computed } from "ccstate";
-import type { AppRoute } from "@ts-rest/core";
 import {
-  zeroConnectorAuthorizeContract,
   zeroConnectorManualGrantContract,
   zeroConnectorOauthStartContract,
-  zeroConnectorSessionsContract,
-  zeroConnectorSessionByIdContract,
   zeroConnectorScopeDiffContract,
   zeroConnectorsByTypeContract,
   zeroConnectorsMainContract,
   zeroConnectorsSearchContract,
 } from "@vm0/api-contracts/contracts/zero-connectors";
-import {
-  connectorAuthMethodIdSchema,
-  connectorTypeSchema,
-  type ConnectorAuthMethodId,
-  type ConnectorType,
+import type {
+  ConnectorAuthMethodId,
+  ConnectorType,
 } from "@vm0/connectors/connectors";
 import {
   getConnectorAuthMethod,
   getConnectorAuthMethodIdsForGrantKind,
 } from "@vm0/connectors/connector-utils";
 import { connectorOauthStates } from "@vm0/db/schema/connector-oauth-state";
-import { connectorSessions } from "@vm0/db/schema/connector-session";
-import { and, eq } from "drizzle-orm";
-import { z } from "zod";
 
-import {
-  authContext$,
-  organizationAuthContext$,
-  requiredAuthContext$,
-} from "../auth/auth-context";
+import { authContext$, organizationAuthContext$ } from "../auth/auth-context";
 import { authRoute } from "../auth/auth-route";
 import { request$ } from "../context/hono";
 import { bodyResultOf, pathParamsOf, queryOf } from "../context/request";
 import { badRequestMessage, notFound } from "../../lib/error";
 import { optionalEnv } from "../../lib/env";
 import { nowDate } from "../../lib/time";
-import { writeDb$, type Db } from "../external/db";
+import { writeDb$ } from "../external/db";
 import {
   connectManualGrantConnector$,
   deleteZeroConnectorLocalState$,
@@ -50,36 +35,13 @@ import {
 } from "../services/zero-connector-data.service";
 import { userConnectorAvailability } from "../services/connector-availability.service";
 import type { RouteEntry } from "../route";
-import {
-  getConnectorOAuthCanonicalRedirectUrl,
-  getConnectorOAuthOrigin,
-} from "./connector-oauth-origin";
-import {
-  buildConnectorOAuthCookieHeader,
-  CONNECTOR_OAUTH_CONTEXT_COOKIE_NAME,
-  CONNECTOR_OAUTH_COOKIE_MAX_AGE_SECONDS,
-  CONNECTOR_OAUTH_PKCE_COOKIE_NAME,
-  CONNECTOR_OAUTH_SESSION_COOKIE_NAME,
-  CONNECTOR_OAUTH_STATE_COOKIE_NAME,
-  connectorOAuthRedirectResponse,
-} from "./connector-oauth-route-state";
+import { getConnectorOAuthOrigin } from "./connector-oauth-origin";
+import { CONNECTOR_OAUTH_COOKIE_MAX_AGE_SECONDS } from "./connector-oauth-route-state";
 import {
   buildResolvedConnectorAuthCodeAuthUrl,
   prepareResolvedConnectorAuthCodeStart,
   resolveConnectorAuthCodeStartMethod,
 } from "./connector-auth-code-start";
-
-const CONNECTOR_SESSION_TTL_SECONDS = 15 * 60;
-const CONNECTOR_SESSION_POLL_INTERVAL_SECONDS = 5;
-const CONNECTOR_SESSION_CODE_LENGTH = 8;
-const CONNECTOR_SESSION_CODE_CHARS = "ABCDEFGHJKMNPQRSTUVWXYZ23456789";
-
-type ConnectorAuthorizeRoute = AppRoute & {
-  readonly pathParams: z.ZodType<{ readonly type: string }>;
-  readonly query: z.ZodType<{ readonly session?: string }>;
-};
-
-const connectorSessionIdSchema = z.uuid();
 
 type ResolvedAuthCodeStartMethod = ReturnType<
   typeof resolveConnectorAuthCodeStartMethod
@@ -106,87 +68,6 @@ function connectorUnavailable(type: string) {
       },
     },
   };
-}
-
-function generateConnectorSessionCode(
-  length: number = CONNECTOR_SESSION_CODE_LENGTH,
-): string {
-  let code = "";
-  for (let i = 0; i < length; i++) {
-    if (i > 0 && i % 4 === 0) {
-      code += "-";
-    }
-    code +=
-      CONNECTOR_SESSION_CODE_CHARS[
-        randomInt(CONNECTOR_SESSION_CODE_CHARS.length)
-      ];
-  }
-  return code;
-}
-
-function jsonResponse(body: unknown, status: number): Response {
-  return new Response(JSON.stringify(body), {
-    status,
-    headers: { "Content-Type": "application/json" },
-  });
-}
-
-function appendConnectorOAuthCookie(
-  response: Response,
-  name: string,
-  value: string | undefined,
-): void {
-  if (!value) {
-    return;
-  }
-  response.headers.append(
-    "Set-Cookie",
-    buildConnectorOAuthCookieHeader(
-      name,
-      value,
-      CONNECTOR_OAUTH_COOKIE_MAX_AGE_SECONDS,
-    ),
-  );
-}
-
-function connectorOAuthStartRedirectResponse(args: {
-  readonly url: string;
-  readonly state: string;
-  readonly codeVerifier?: string;
-  readonly oauthContext?: string;
-  readonly session?: string;
-}): Response {
-  const response = connectorOAuthRedirectResponse(args.url);
-  appendConnectorOAuthCookie(
-    response,
-    CONNECTOR_OAUTH_STATE_COOKIE_NAME,
-    args.state,
-  );
-  appendConnectorOAuthCookie(
-    response,
-    CONNECTOR_OAUTH_PKCE_COOKIE_NAME,
-    args.codeVerifier,
-  );
-  appendConnectorOAuthCookie(
-    response,
-    CONNECTOR_OAUTH_CONTEXT_COOKIE_NAME,
-    args.oauthContext,
-  );
-  appendConnectorOAuthCookie(
-    response,
-    CONNECTOR_OAUTH_SESSION_COOKIE_NAME,
-    args.session,
-  );
-  return response;
-}
-
-function connectorMissingAuthCodeGrantResponse(type: string) {
-  return jsonResponse(
-    {
-      error: `${type} connector does not use an auth-code grant`,
-    },
-    400,
-  );
 }
 
 function connectorTypeHasAuthCodeGrant(type: ConnectorType): boolean {
@@ -220,41 +101,6 @@ function connectorAuthCodeStartErrorMessage(
   }
 }
 
-function connectorAuthorizeAuthCodeStartErrorResponse(
-  type: ConnectorType,
-  reason:
-    | "invalid_session"
-    | "missing_auth_code_grant"
-    | "missing_auth_method"
-    | "wrong_grant_kind",
-) {
-  if (reason === "invalid_session") {
-    return jsonResponse({ error: "Invalid connector session" }, 400);
-  }
-  if (!connectorTypeHasAuthCodeGrant(type)) {
-    return connectorMissingAuthCodeGrantResponse(type);
-  }
-  if (reason === "missing_auth_method" || reason === "wrong_grant_kind") {
-    return jsonResponse(
-      { error: "Invalid connector session auth method" },
-      500,
-    );
-  }
-  return connectorMissingAuthCodeGrantResponse(type);
-}
-
-function connectorAuthorizeMissingOrgResponse() {
-  return jsonResponse(
-    {
-      error: {
-        message: "Explicit org context required — ensure active org in session",
-        code: "BAD_REQUEST",
-      },
-    },
-    400,
-  );
-}
-
 function resolveRequestedAuthCodeStartMethod(
   type: ConnectorType,
   authMethod: ConnectorAuthMethodId,
@@ -264,47 +110,6 @@ function resolveRequestedAuthCodeStartMethod(
     return result;
   }
   return { ok: false, reason: "wrong_grant_kind" };
-}
-
-async function resolveSessionAuthCodeStartMethod(args: {
-  readonly writeDb: Db;
-  readonly sessionId: string;
-  readonly type: ConnectorType;
-  readonly userId: string;
-}): Promise<
-  | ResolvedAuthCodeStartMethod
-  | { readonly ok: false; readonly reason: "invalid_session" }
-> {
-  const [session] = await args.writeDb
-    .select({
-      type: connectorSessions.type,
-      authMethod: connectorSessions.authMethod,
-      userId: connectorSessions.userId,
-      status: connectorSessions.status,
-      expiresAt: connectorSessions.expiresAt,
-    })
-    .from(connectorSessions)
-    .where(eq(connectorSessions.id, args.sessionId))
-    .limit(1);
-
-  if (
-    !session ||
-    session.type !== args.type ||
-    session.userId !== args.userId ||
-    session.status !== "pending" ||
-    session.expiresAt <= nowDate()
-  ) {
-    return { ok: false, reason: "invalid_session" };
-  }
-
-  const authMethodResult = connectorAuthMethodIdSchema.safeParse(
-    session.authMethod,
-  );
-  if (!authMethodResult.success) {
-    return { ok: false, reason: "missing_auth_method" };
-  }
-
-  return resolveRequestedAuthCodeStartMethod(args.type, authMethodResult.data);
 }
 
 function internalServerError(message: string) {
@@ -458,140 +263,6 @@ const connectManualGrantConnectorInner$ = command(
   },
 );
 
-export function createAuthorizeConnectorInner(route: ConnectorAuthorizeRoute) {
-  return command(async ({ get, set }, signal: AbortSignal) => {
-    const params = get(pathParamsOf(route));
-    const query = get(queryOf(route));
-    const request = get(request$).raw;
-    const canonicalRedirectUrl = getConnectorOAuthCanonicalRedirectUrl(request);
-    if (canonicalRedirectUrl) {
-      return connectorOAuthRedirectResponse(canonicalRedirectUrl);
-    }
-    const origin = getConnectorOAuthOrigin(request);
-    const requestUrl = new URL(request.url);
-
-    const typeResult = connectorTypeSchema.safeParse(params.type);
-    if (!typeResult.success) {
-      return jsonResponse(
-        { error: `Unknown connector type: ${params.type}` },
-        400,
-      );
-    }
-    const type = typeResult.data;
-
-    const sessionResult = query.session
-      ? connectorSessionIdSchema.safeParse(query.session)
-      : undefined;
-    if (!sessionResult?.success) {
-      return jsonResponse({ error: "Invalid connector session" }, 400);
-    }
-    const sessionId = sessionResult.data;
-
-    const auth = await set(
-      requiredAuthContext$,
-      { requireOrganization: true },
-      signal,
-    );
-    signal.throwIfAborted();
-    if ("status" in auth) {
-      if (auth.status === 401) {
-        const loginUrl = new URL("/sign-in", origin);
-        const authorizeUrl = new URL(
-          `${requestUrl.pathname}${requestUrl.search}`,
-          origin,
-        );
-        loginUrl.searchParams.set("redirect_url", authorizeUrl.toString());
-        return connectorOAuthRedirectResponse(loginUrl.toString());
-      }
-      return jsonResponse(auth.body, auth.status);
-    }
-
-    if (!auth.orgId) {
-      return connectorAuthorizeMissingOrgResponse();
-    }
-
-    const availability = await get(
-      userConnectorAvailability(auth.orgId, auth.userId),
-    );
-    signal.throwIfAborted();
-    const writeDb = set(writeDb$);
-    const authCodeStartType = await resolveSessionAuthCodeStartMethod({
-      writeDb,
-      sessionId,
-      type,
-      userId: auth.userId,
-    });
-    signal.throwIfAborted();
-    if (!authCodeStartType.ok) {
-      return connectorAuthorizeAuthCodeStartErrorResponse(
-        type,
-        authCodeStartType.reason,
-      );
-    }
-    if (
-      !availability.isAuthMethodAvailable(
-        authCodeStartType.type,
-        authCodeStartType.authMethod,
-      )
-    ) {
-      return jsonResponse({ error: `${type} connector is not available` }, 403);
-    }
-
-    const prepared = prepareResolvedConnectorAuthCodeStart({
-      type: authCodeStartType.type,
-      authMethod: authCodeStartType.authMethod,
-      origin,
-      readEnv: optionalEnv,
-    });
-    if (!prepared.ok) {
-      return jsonResponse({ error: `${type} OAuth not configured` }, 500);
-    }
-    const authResult = await buildResolvedConnectorAuthCodeAuthUrl({
-      type: authCodeStartType.type,
-      authMethod: authCodeStartType.authMethod,
-      authClient: prepared.authClient,
-      redirectUri: prepared.redirectUri,
-      state: prepared.state,
-    });
-    signal.throwIfAborted();
-
-    await set(
-      deleteZeroConnectorLocalState$,
-      { orgId: auth.orgId, userId: auth.userId, type },
-      signal,
-    );
-    signal.throwIfAborted();
-
-    await writeDb.insert(connectorOauthStates).values({
-      state: prepared.state,
-      type: authCodeStartType.type,
-      authMethod: authCodeStartType.authMethod,
-      userId: auth.userId,
-      orgId: auth.orgId,
-      redirectUri: prepared.redirectUri,
-      sessionId,
-      codeVerifier: authResult.codeVerifier,
-      oauthContext: authResult.oauthContext,
-      expiresAt: new Date(
-        nowDate().getTime() + CONNECTOR_OAUTH_COOKIE_MAX_AGE_SECONDS * 1000,
-      ),
-    });
-    signal.throwIfAborted();
-
-    return connectorOAuthStartRedirectResponse({
-      url: authResult.url,
-      state: prepared.state,
-      codeVerifier: authResult.codeVerifier,
-      oauthContext: authResult.oauthContext,
-      session: sessionId,
-    });
-  });
-}
-
-const authorizeConnectorInner$ = createAuthorizeConnectorInner(
-  zeroConnectorAuthorizeContract.authorize,
-);
-
 const startConnectorOauthInner$ = command(
   async ({ get, set }, signal: AbortSignal) => {
     const params = get(pathParamsOf(zeroConnectorOauthStartContract.start));
@@ -690,132 +361,6 @@ const startConnectorOauthInner$ = command(
   },
 );
 
-const createConnectorSessionInner$ = command(
-  async ({ get, set }, signal: AbortSignal) => {
-    const auth = get(organizationAuthContext$);
-    const params = get(pathParamsOf(zeroConnectorSessionsContract.create));
-    const bodyResult = await get(
-      bodyResultOf(zeroConnectorSessionsContract.create),
-    );
-    signal.throwIfAborted();
-    if (!bodyResult.ok) {
-      return bodyResult.response;
-    }
-
-    const authCodeStartType = resolveRequestedAuthCodeStartMethod(
-      params.type,
-      bodyResult.data.authMethod,
-    );
-    if (!authCodeStartType.ok) {
-      return badRequestMessage(
-        connectorAuthCodeStartErrorMessage(
-          params.type,
-          bodyResult.data.authMethod,
-          authCodeStartType.reason,
-        ),
-      );
-    }
-
-    const availability = await get(
-      userConnectorAvailability(auth.orgId, auth.userId),
-    );
-    signal.throwIfAborted();
-    if (
-      !availability.isAuthMethodAvailable(
-        authCodeStartType.type,
-        authCodeStartType.authMethod,
-      )
-    ) {
-      return connectorUnavailable(params.type);
-    }
-
-    const code = generateConnectorSessionCode();
-    const expiresAt = new Date(
-      nowDate().getTime() + CONNECTOR_SESSION_TTL_SECONDS * 1000,
-    );
-    const writeDb = set(writeDb$);
-
-    const [session] = await writeDb
-      .insert(connectorSessions)
-      .values({
-        code,
-        type: authCodeStartType.type,
-        authMethod: authCodeStartType.authMethod,
-        userId: auth.userId,
-        status: "pending",
-        expiresAt,
-      })
-      .returning({ id: connectorSessions.id });
-    signal.throwIfAborted();
-
-    if (!session) {
-      throw new Error("Failed to create connector session");
-    }
-
-    return {
-      status: 200 as const,
-      body: {
-        id: session.id,
-        code,
-        type: authCodeStartType.type,
-        status: "pending" as const,
-        verificationUrl: `/api/connectors/${authCodeStartType.type}/authorize?session=${session.id}`,
-        expiresIn: CONNECTOR_SESSION_TTL_SECONDS,
-        interval: CONNECTOR_SESSION_POLL_INTERVAL_SECONDS,
-      },
-    };
-  },
-);
-
-const getConnectorSessionByIdInner$ = command(
-  async ({ get, set }, signal: AbortSignal) => {
-    const auth = get(authContext$);
-    const params = get(pathParamsOf(zeroConnectorSessionByIdContract.get));
-    const writeDb = set(writeDb$);
-
-    const [session] = await writeDb
-      .select()
-      .from(connectorSessions)
-      .where(
-        and(
-          eq(connectorSessions.id, params.sessionId),
-          eq(connectorSessions.type, params.type),
-          eq(connectorSessions.userId, auth.userId),
-        ),
-      )
-      .limit(1);
-    signal.throwIfAborted();
-
-    if (!session) {
-      return notFound("Connector session not found");
-    }
-
-    if (session.status === "pending" && nowDate() > session.expiresAt) {
-      await writeDb
-        .update(connectorSessions)
-        .set({ status: "expired" })
-        .where(eq(connectorSessions.id, session.id));
-      signal.throwIfAborted();
-
-      return {
-        status: 200 as const,
-        body: {
-          status: "expired" as const,
-          errorMessage: "Session has expired",
-        },
-      };
-    }
-
-    return {
-      status: 200 as const,
-      body: {
-        status: session.status,
-        errorMessage: session.errorMessage,
-      },
-    };
-  },
-);
-
 export const zeroConnectorsRoutes: readonly RouteEntry[] = [
   {
     route: zeroConnectorManualGrantContract.connect,
@@ -834,20 +379,8 @@ export const zeroConnectorsRoutes: readonly RouteEntry[] = [
     handler: authRoute(connectorReadAuth, getScopeDiffInner$),
   },
   {
-    route: zeroConnectorAuthorizeContract.authorize,
-    handler: authorizeConnectorInner$,
-  },
-  {
     route: zeroConnectorOauthStartContract.start,
     handler: authRoute(connectorWriteAuth, startConnectorOauthInner$),
-  },
-  {
-    route: zeroConnectorSessionByIdContract.get,
-    handler: authRoute({}, getConnectorSessionByIdInner$),
-  },
-  {
-    route: zeroConnectorSessionsContract.create,
-    handler: authRoute(connectorWriteAuth, createConnectorSessionInner$),
   },
   {
     route: zeroConnectorsByTypeContract.get,
