+++ turbo/apps/api/src/signals/services/webhooks-clerk-cleanup.service.ts
@@ -8,7 +8,6 @@ import { agentRuns } from "@vm0/db/schema/agent-run";
 import { cliTokens } from "@vm0/db/schema/cli-tokens";
 import { composeJobs } from "@vm0/db/schema/compose-job";
 import { connectorOauthDeviceAuthorizationSessions } from "@vm0/db/schema/connector-oauth-device-authorization-session";
-import { connectorSessions } from "@vm0/db/schema/connector-session";
 import { connectors } from "@vm0/db/schema/connector";
 import { deviceCodes } from "@vm0/db/schema/device-codes";
 import { exportJobs } from "@vm0/db/schema/export-job";
@@ -544,9 +543,6 @@ async function deleteUserData(db: Db, userId: string): Promise<void> {
     .where(eq(zeroAgentSchedules.userId, userId));
   await db.delete(cliTokens).where(eq(cliTokens.userId, userId));
   await db.delete(composeJobs).where(eq(composeJobs.userId, userId));
-  await db
-    .delete(connectorSessions)
-    .where(eq(connectorSessions.userId, userId));
   await db
     .delete(connectorOauthDeviceAuthorizationSessions)
     .where(eq(connectorOauthDeviceAuthorizationSessions.userId, userId));
