+++ turbo/apps/web/__tests__/proxy.sandbox-token.test.ts
@@ -255,36 +255,6 @@ describe("proxy middleware: sandbox token handling", () => {
     ).toBe("preview-secret");
   });
 
-  it("should preserve web origin headers for API backend OAuth authorize routes", async () => {
-    const request = new NextRequest(
-      "https://www.vm0.ai/api/zero/connectors/google-calendar/authorize",
-      {
-        headers: {
-          cookie: "__session=opaque",
-        },
-      },
-    );
-
-    const response = await middleware(request, createMockEvent());
-    if (!response) {
-      throw new Error("Expected middleware response");
-    }
-
-    expect(capturedClerkRequest).toBeUndefined();
-    expect(response.headers.get("x-middleware-request-cookie")).toBe(
-      "__session=opaque",
-    );
-    expect(response.headers.get("x-middleware-request-x-forwarded-host")).toBe(
-      "www.vm0.ai",
-    );
-    expect(response.headers.get("x-middleware-request-x-forwarded-proto")).toBe(
-      "https",
-    );
-    expect(response.headers.get("x-middleware-request-x-vm0-web-origin")).toBe(
-      "https://www.vm0.ai",
-    );
-  });
-
   it("should preserve web origin headers for API backend OAuth start routes", async () => {
     const request = new NextRequest(
       "https://www.vm0.ai/api/zero/connectors/google-calendar/oauth/start",
