+++ turbo/apps/web/__tests__/security-headers.test.ts
@@ -490,13 +490,6 @@ const CRON_TELEGRAM_CLEANUP_NEXT_NEGATIVE_PATHS = [
   "/api/cron/telegram-cleanup/extra",
   "/api/cron",
 ] as const;
-const CONNECTORS_AUTHORIZE_REWRITE_SOURCE = "/api/connectors/:type/authorize";
-const CONNECTORS_AUTHORIZE_PATH = "/api/connectors/github/authorize";
-const CONNECTORS_AUTHORIZE_NEXT_NEGATIVE_PATHS = [
-  "/api/connectors/github/authorize/extra",
-  "/api/connectors/authorize",
-  "/api/connectors/github/authorizes",
-] as const;
 const CONNECTORS_CALLBACK_REWRITE_SOURCE = "/api/connectors/:type/callback";
 const CONNECTORS_CALLBACK_PATH = "/api/connectors/github/callback";
 const CONNECTORS_CALLBACK_NEXT_NEGATIVE_PATHS = [
@@ -1640,14 +1633,6 @@ const VOICE_IO_TTS_NEXT_NEGATIVE_PATHS = [
   "/api/zero/voice-io/speech",
   "/api/zero/voice-io/stt",
 ] as const;
-const ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE =
-  "/api/zero/connectors/:type/authorize";
-const ZERO_CONNECTORS_AUTHORIZE_PATH = "/api/zero/connectors/github/authorize";
-const ZERO_CONNECTORS_AUTHORIZE_NEXT_NEGATIVE_PATHS = [
-  "/api/zero/connectors/github/authorize/extra",
-  "/api/zero/connectors/authorize",
-  "/api/zero/connectors/github/callback",
-] as const;
 const ZERO_CONNECTORS_LIST_REWRITE_SOURCE = "/api/zero/connectors";
 const ZERO_CONNECTORS_LIST_PATH = "/api/zero/connectors";
 const ZERO_CONNECTORS_LIST_NEXT_NEGATIVE_PATHS = [
@@ -1687,22 +1672,6 @@ const ZERO_CONNECTORS_SCOPE_DIFF_NEXT_NEGATIVE_PATHS = [
   "/api/zero/connectors/github/scope-diff/extra",
   "/api/zero/connectors/scope-diff",
   "/api/zero/connectors/github/scope",
-] as const;
-const ZERO_CONNECTORS_SESSIONS_REWRITE_SOURCE =
-  "/api/zero/connectors/:type/sessions";
-const ZERO_CONNECTORS_SESSIONS_PATH = "/api/zero/connectors/github/sessions";
-const ZERO_CONNECTORS_SESSIONS_NEXT_NEGATIVE_PATHS = [
-  "/api/zero/connectors/github/sessions/00000000-0000-0000-0000-000000000000",
-  "/api/zero/connectors/sessions",
-  "/api/zero/connectors/github/session",
-] as const;
-const ZERO_CONNECTORS_SESSION_BY_ID_REWRITE_SOURCE =
-  "/api/zero/connectors/:type/sessions/:sessionId([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})";
-const ZERO_CONNECTORS_SESSION_BY_ID_PATH = `/api/zero/connectors/github/sessions/${ZERO_CONNECTOR_SESSION_ID}`;
-const ZERO_CONNECTORS_SESSION_BY_ID_NEXT_NEGATIVE_PATHS = [
-  "/api/zero/connectors/github/sessions/not-a-uuid",
-  `/api/zero/connectors/github/sessions/${ZERO_CONNECTOR_SESSION_ID}/extra`,
-  `/api/zero/connectors/sessions/${ZERO_CONNECTOR_SESSION_ID}`,
 ] as const;
 const ZERO_CONNECTORS_OAUTH_DEVICE_AUTHORIZATION_SESSIONS_REWRITE_SOURCE =
   "/api/zero/connectors/:type/oauth/device/sessions";
@@ -2095,11 +2064,6 @@ describe("API backend rewrites", () => {
           destination: "https://api.example.test/api/cron/telegram-cleanup",
         },
         {
-          source: CONNECTORS_AUTHORIZE_REWRITE_SOURCE,
-          destination:
-            "https://api.example.test/api/connectors/:type/authorize",
-        },
-        {
           source: CONNECTORS_CALLBACK_REWRITE_SOURCE,
           destination: "https://api.example.test/api/connectors/:type/callback",
         },
@@ -2475,11 +2439,6 @@ describe("API backend rewrites", () => {
           source: ZERO_CONNECTORS_MANUAL_GRANT_REWRITE_SOURCE,
           destination:
             "https://api.example.test/api/zero/connectors/:type/manual-grant",
-        },
-        {
-          source: ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE,
-          destination:
-            "https://api.example.test/api/zero/connectors/:type/authorize",
         },
         {
           source: ZERO_SLACK_CHANNELS_REWRITE_SOURCE,
@@ -3794,31 +3753,6 @@ describe("API backend rewrites", () => {
     }
   });
 
-  it("should match only the connector authorize rewrite", async () => {
-    vi.stubEnv("VM0_API_BACKEND_URL", "https://api.example.test");
-
-    const rewrites = await getBeforeFileRewrites();
-    const rewrite = rewrites.find((entry) => {
-      return entry.source === CONNECTORS_AUTHORIZE_REWRITE_SOURCE;
-    });
-    expect(rewrite).toStrictEqual({
-      source: CONNECTORS_AUTHORIZE_REWRITE_SOURCE,
-      destination: "https://api.example.test/api/connectors/:type/authorize",
-    });
-
-    const matcher = getPathMatch(CONNECTORS_AUTHORIZE_REWRITE_SOURCE, {
-      removeUnnamedParams: true,
-      strict: true,
-    });
-
-    expect(matcher(CONNECTORS_AUTHORIZE_PATH)).toStrictEqual({
-      type: "github",
-    });
-    for (const pathname of CONNECTORS_AUTHORIZE_NEXT_NEGATIVE_PATHS) {
-      expect(matcher(pathname)).toBe(false);
-    }
-  });
-
   it("should match only the connector callback rewrite", async () => {
     vi.stubEnv("VM0_API_BACKEND_URL", "https://api.example.test");
 
@@ -4289,59 +4223,6 @@ describe("API backend rewrites", () => {
     }
   });
 
-  it("should match only the zero connector sessions rewrite", async () => {
-    vi.stubEnv("VM0_API_BACKEND_URL", "https://api.example.test");
-
-    const rewrites = await getBeforeFileRewrites();
-    const rewrite = rewrites.find((entry) => {
-      return entry.source === ZERO_CONNECTORS_SESSIONS_REWRITE_SOURCE;
-    });
-    expect(rewrite).toStrictEqual({
-      source: ZERO_CONNECTORS_SESSIONS_REWRITE_SOURCE,
-      destination:
-        "https://api.example.test/api/zero/connectors/:type/sessions",
-    });
-
-    const matcher = getPathMatch(ZERO_CONNECTORS_SESSIONS_REWRITE_SOURCE, {
-      removeUnnamedParams: true,
-      strict: true,
-    });
-
-    expect(matcher(ZERO_CONNECTORS_SESSIONS_PATH)).toStrictEqual({
-      type: "github",
-    });
-    for (const pathname of ZERO_CONNECTORS_SESSIONS_NEXT_NEGATIVE_PATHS) {
-      expect(matcher(pathname)).toBe(false);
-    }
-  });
-
-  it("should match only UUID-shaped zero connector session polling rewrites", async () => {
-    vi.stubEnv("VM0_API_BACKEND_URL", "https://api.example.test");
-
-    const rewrites = await getBeforeFileRewrites();
-    const rewrite = rewrites.find((entry) => {
-      return entry.source === ZERO_CONNECTORS_SESSION_BY_ID_REWRITE_SOURCE;
-    });
-    expect(rewrite).toStrictEqual({
-      source: ZERO_CONNECTORS_SESSION_BY_ID_REWRITE_SOURCE,
-      destination:
-        "https://api.example.test/api/zero/connectors/:type/sessions/:sessionId",
-    });
-
-    const matcher = getPathMatch(ZERO_CONNECTORS_SESSION_BY_ID_REWRITE_SOURCE, {
-      removeUnnamedParams: true,
-      strict: true,
-    });
-
-    expect(matcher(ZERO_CONNECTORS_SESSION_BY_ID_PATH)).toStrictEqual({
-      type: "github",
-      sessionId: ZERO_CONNECTOR_SESSION_ID,
-    });
-    for (const pathname of ZERO_CONNECTORS_SESSION_BY_ID_NEXT_NEGATIVE_PATHS) {
-      expect(matcher(pathname)).toBe(false);
-    }
-  });
-
   it("should match only the zero connector OAuth device session rewrite", async () => {
     vi.stubEnv("VM0_API_BACKEND_URL", "https://api.example.test");
 
@@ -4524,32 +4405,6 @@ describe("API backend rewrites", () => {
 
     expect(matcher(ZERO_SLACK_INTERACTIVE_PATH)).toStrictEqual({});
     for (const pathname of ZERO_SLACK_INTERACTIVE_NEXT_NEGATIVE_PATHS) {
-      expect(matcher(pathname)).toBe(false);
-    }
-  });
-
-  it("should match only the zero connector authorize rewrite", async () => {
-    vi.stubEnv("VM0_API_BACKEND_URL", "https://api.example.test");
-
-    const rewrites = await getBeforeFileRewrites();
-    const rewrite = rewrites.find((entry) => {
-      return entry.source === ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE;
-    });
-    expect(rewrite).toStrictEqual({
-      source: ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE,
-      destination:
-        "https://api.example.test/api/zero/connectors/:type/authorize",
-    });
-
-    const matcher = getPathMatch(ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE, {
-      removeUnnamedParams: true,
-      strict: true,
-    });
-
-    expect(matcher(ZERO_CONNECTORS_AUTHORIZE_PATH)).toStrictEqual({
-      type: "github",
-    });
-    for (const pathname of ZERO_CONNECTORS_AUTHORIZE_NEXT_NEGATIVE_PATHS) {
       expect(matcher(pathname)).toBe(false);
     }
   });
