+++ turbo/apps/web/api-backend-rewrites.js
@@ -126,8 +126,6 @@ const AGENT_RUN_TELEMETRY_NETWORK_REWRITE_SOURCE = `/api/agent/runs/:id(${UUID_P
 const AGENT_RUN_TELEMETRY_NETWORK_PATH_RE = new RegExp(
   `^/api/agent/runs/${UUID_PATH_SEGMENT_PATTERN}/telemetry/network$`,
 );
-const CONNECTORS_AUTHORIZE_REWRITE_SOURCE = "/api/connectors/:type/authorize";
-const CONNECTORS_AUTHORIZE_PATH_RE = /^\/api\/connectors\/[^/]+\/authorize$/;
 const AGENT_RUN_TELEMETRY_SYSTEM_LOG_REWRITE_SOURCE = `/api/agent/runs/:id(${UUID_PATH_SEGMENT_PATTERN})/telemetry/system-log`;
 const AGENT_RUN_TELEMETRY_SYSTEM_LOG_PATH_RE = new RegExp(
   `^/api/agent/runs/${UUID_PATH_SEGMENT_PATTERN}/telemetry/system-log$`,
@@ -265,10 +263,6 @@ const ZERO_COMPUTER_USE_HOSTS_START_REWRITE_SOURCE =
   "/api/zero/computer-use/hosts/start";
 const ZERO_COMPUTER_USE_WRITE_COMMANDS_REWRITE_SOURCE =
   "/api/zero/computer-use/write-commands";
-const ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE =
-  "/api/zero/connectors/:type/authorize";
-const ZERO_CONNECTORS_AUTHORIZE_PATH_RE =
-  /^\/api\/zero\/connectors\/[^/]+\/authorize$/;
 const ZERO_CONNECTORS_LIST_REWRITE_SOURCE = "/api/zero/connectors";
 const ZERO_CONNECTORS_SEARCH_REWRITE_SOURCE = "/api/zero/connectors/search";
 const ZERO_CONNECTORS_BY_TYPE_REWRITE_SOURCE =
@@ -283,14 +277,6 @@ const ZERO_CONNECTORS_SCOPE_DIFF_REWRITE_SOURCE =
   "/api/zero/connectors/:type/scope-diff";
 const ZERO_CONNECTORS_SCOPE_DIFF_PATH_RE =
   /^\/api\/zero\/connectors\/[^/]+\/scope-diff$/;
-const ZERO_CONNECTORS_SESSIONS_REWRITE_SOURCE =
-  "/api/zero/connectors/:type/sessions";
-const ZERO_CONNECTORS_SESSIONS_PATH_RE =
-  /^\/api\/zero\/connectors\/[^/]+\/sessions$/;
-const ZERO_CONNECTORS_SESSION_BY_ID_REWRITE_SOURCE = `/api/zero/connectors/:type/sessions/:sessionId(${UUID_PATH_SEGMENT_PATTERN})`;
-const ZERO_CONNECTORS_SESSION_BY_ID_PATH_RE = new RegExp(
-  `^/api/zero/connectors/[^/]+/sessions/${UUID_PATH_SEGMENT_PATTERN}$`,
-);
 const ZERO_CONNECTORS_OAUTH_START_REWRITE_SOURCE =
   "/api/zero/connectors/:type/oauth/start";
 const ZERO_CONNECTORS_OAUTH_START_PATH_RE =
@@ -464,11 +450,6 @@ export const API_BACKEND_REWRITES = [
   ],
   ["/api/cron/sync-skills", "/api/cron/sync-skills"],
   ["/api/cron/telegram-cleanup", "/api/cron/telegram-cleanup"],
-  [
-    CONNECTORS_AUTHORIZE_REWRITE_SOURCE,
-    "/api/connectors/:type/authorize",
-    CONNECTORS_AUTHORIZE_PATH_RE,
-  ],
   [
     CONNECTORS_CALLBACK_REWRITE_SOURCE,
     "/api/connectors/:type/callback",
@@ -724,21 +705,6 @@ export const API_BACKEND_REWRITES = [
     ZERO_CONNECTORS_SCOPE_DIFF_REWRITE_SOURCE,
     "/api/zero/connectors/:type/scope-diff",
     ZERO_CONNECTORS_SCOPE_DIFF_PATH_RE,
-  ],
-  [
-    ZERO_CONNECTORS_SESSIONS_REWRITE_SOURCE,
-    "/api/zero/connectors/:type/sessions",
-    ZERO_CONNECTORS_SESSIONS_PATH_RE,
-  ],
-  [
-    ZERO_CONNECTORS_SESSION_BY_ID_REWRITE_SOURCE,
-    "/api/zero/connectors/:type/sessions/:sessionId",
-    ZERO_CONNECTORS_SESSION_BY_ID_PATH_RE,
-  ],
-  [
-    ZERO_CONNECTORS_AUTHORIZE_REWRITE_SOURCE,
-    "/api/zero/connectors/:type/authorize",
-    ZERO_CONNECTORS_AUTHORIZE_PATH_RE,
   ],
   [
     ZERO_CONNECTORS_OAUTH_START_REWRITE_SOURCE,
@@ -1156,9 +1122,7 @@ export function matchesApiBackendRewritePath(pathname) {
 
 export function matchesConnectorOAuthRewritePath(pathname) {
   return (
-    CONNECTORS_AUTHORIZE_PATH_RE.test(pathname) ||
     CONNECTORS_CALLBACK_PATH_RE.test(pathname) ||
-    ZERO_CONNECTORS_AUTHORIZE_PATH_RE.test(pathname) ||
     ZERO_CONNECTORS_OAUTH_START_PATH_RE.test(pathname)
   );
 }
