+++ turbo/packages/api-contracts/src/contracts/zero-connectors.ts
@@ -12,8 +12,6 @@ import {
   connectorOauthStartResponseSchema,
   connectorListResponseSchema,
   connectorResponseSchema,
-  connectorSessionResponseSchema,
-  connectorSessionStatusResponseSchema,
   scopeDiffResponseSchema,
 } from "./connector-schemas";
 
@@ -87,29 +85,6 @@ export const zeroConnectorScopeDiffContract = c.router({
       404: apiErrorSchema,
     },
     summary: "Get scope diff for a connector",
-  },
-});
-
-/**
- * Zero contract for GET /api/zero/connectors/:type/authorize
- * Browser OAuth redirect endpoint. The API handler returns raw Response
- * redirects so Set-Cookie and Location headers are preserved.
- */
-export const zeroConnectorAuthorizeContract = c.router({
-  authorize: {
-    method: "GET",
-    path: "/api/zero/connectors/:type/authorize",
-    headers: authHeadersSchema,
-    pathParams: z.object({ type: z.string() }),
-    query: z.object({ session: z.string().optional() }),
-    responses: {
-      307: c.noBody(),
-      400: z.object({ error: z.string() }),
-      401: c.noBody(),
-      403: z.object({ error: z.string() }),
-      500: z.object({ error: z.string() }),
-    },
-    summary: "Start connector OAuth authorization (zero proxy)",
   },
 });
 
@@ -229,62 +204,12 @@ export const zeroConnectorsSearchContract = c.router({
   },
 });
 
-/**
- * Zero contract for POST /api/zero/connectors/:type/sessions.
- * Creates an auth-code connector handoff session.
- */
-export const zeroConnectorSessionsContract = c.router({
-  create: {
-    method: "POST",
-    path: "/api/zero/connectors/:type/sessions",
-    headers: authHeadersSchema,
-    pathParams: z.object({ type: connectorTypeSchema }),
-    body: z.object({ authMethod: connectorAuthMethodIdSchema }),
-    responses: {
-      200: connectorSessionResponseSchema,
-      400: apiErrorSchema,
-      401: apiErrorSchema,
-      403: apiErrorSchema,
-    },
-    summary: "Create connector session for auth-code handoff",
-  },
-});
-
-/**
- * Zero contract for GET /api/zero/connectors/:type/sessions/:sessionId
- * Proxies to GET /api/connectors/:type/sessions/:sessionId (poll session)
- */
-export const zeroConnectorSessionByIdContract = c.router({
-  get: {
-    method: "GET",
-    path: "/api/zero/connectors/:type/sessions/:sessionId",
-    headers: authHeadersSchema,
-    pathParams: z.object({
-      type: connectorTypeSchema,
-      sessionId: z.uuid(),
-    }),
-    responses: {
-      200: connectorSessionStatusResponseSchema,
-      400: apiErrorSchema,
-      401: apiErrorSchema,
-      404: apiErrorSchema,
-    },
-    summary: "Get connector session status (zero proxy)",
-  },
-});
-
 export type ZeroConnectorsMainContract = typeof zeroConnectorsMainContract;
 export type ZeroConnectorsByTypeContract = typeof zeroConnectorsByTypeContract;
 export type ZeroConnectorScopeDiffContract =
   typeof zeroConnectorScopeDiffContract;
-export type ZeroConnectorAuthorizeContract =
-  typeof zeroConnectorAuthorizeContract;
 export type ZeroConnectorManualGrantContract =
   typeof zeroConnectorManualGrantContract;
 export type ZeroConnectorOauthDeviceAuthSessionContract =
   typeof zeroConnectorOauthDeviceAuthSessionContract;
 export type ZeroConnectorsSearchContract = typeof zeroConnectorsSearchContract;
-export type ZeroConnectorSessionsContract =
-  typeof zeroConnectorSessionsContract;
-export type ZeroConnectorSessionByIdContract =
-  typeof zeroConnectorSessionByIdContract;
