+++ turbo/packages/core/src/index.ts
@@ -225,11 +225,8 @@ export {
   getConnectorEnvNamesForSecret,
   getConnectorTypeForSecretName,
   isGoogleOAuthConnector,
-  connectorSessionStatusSchema,
   connectorResponseSchema,
   connectorListResponseSchema,
-  connectorSessionResponseSchema,
-  connectorSessionStatusResponseSchema,
   scopeDiffResponseSchema,
   parseBasicAuthTemplates,
   replaceBasicAuthTemplates,
@@ -313,8 +310,6 @@ export {
   zeroConnectorScopeDiffContract,
   zeroConnectorManualGrantContract,
   zeroConnectorsSearchContract,
-  zeroConnectorSessionsContract,
-  zeroConnectorSessionByIdContract,
   zeroOrgContract,
   zeroOrgLeaveContract,
   zeroOrgDeleteContract,
@@ -573,11 +568,8 @@ export {
   type ConnectorAuthMethodId,
   type ScopeDiff,
   type ScopeDiffResponse,
-  type ConnectorSessionStatus,
   type ConnectorResponse,
   type ConnectorListResponse,
-  type ConnectorSessionResponse,
-  type ConnectorSessionStatusResponse,
   type FirewallConfig,
   type ExpandedFirewallConfig,
   type FirewallApi,
@@ -640,8 +632,6 @@ export {
   type ZeroConnectorsByTypeContract,
   type ZeroConnectorScopeDiffContract,
   type ZeroConnectorsSearchContract,
-  type ZeroConnectorSessionsContract,
-  type ZeroConnectorSessionByIdContract,
   type ZeroOrgContract,
   type ZeroOrgLeaveContract,
   type ZeroOrgDeleteContract,
